% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare_utilities.R
\name{csquare_utils}
\alias{csquare_utils}
\alias{csquare_area}
\alias{csquare_lat}
\alias{csquare_lon}
\title{Get information related to a C-Square}
\usage{
csquare_area(csquare)

csquare_lat(csquare)

csquare_lon(csquare)
}
\arguments{
\item{csquare}{the name of a 0.05 resolution C-Square.}
}
\value{
A vector of numeric values: latitudes, longitudes or areas.
}
\description{
Extract the surface area, latitude or longitude of a 0.05 resolution
C-Square.
}
\examples{
csquare_area("1501:370:370:1")
csquare_lat("1501:370:370:1")
csquare_lon("1501:370:370:1")

}
