% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metier_lookup.R
\docType{data}
\name{metier_lookup}
\alias{metier_lookup}
\title{Lookup table to aggregate metier level 6 gear groupings to Benthis
gear groupings}
\format{Data frame with containing 17 columns:
\tabular{ll}{
\code{LE_MET_level6} \tab Metier level 6 gear code\cr
\code{Benthis_metiers} \tab Benthis metier used to define bottom fishing pressure\cr
\code{Metier_level5} \tab Metier level 5 gear codes\cr
\code{Metier_level4} \tab Metier level 5 gear codes\cr
\code{JNCC_grouping} \tab JNCC gear groupings\cr
\code{Fishing_category} \tab Fishing category\cr
\code{Fishing_category_FO} \tab Fishing category used in ICES fishery overview reports\cr
\code{Description} \tab Text description of the gear code\cr
}}
\source{
Reference to ices WGSFD, Benthis, ...
}
\usage{
metier_lookup
}
\description{
A table.
}
\details{
A lookup table linking Metier level 6 codes to the benthis gear groupings
developed at the ICES WGSFD. The lookup table contains other gear groupings
used in ICES outputs.
}
\examples{
head(metier_lookup)
}
\seealso{
\code{\link{sfdSAR-package}} gives an overview of the package.
}
