% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gear_width.R
\name{predict_gear_width}
\alias{predict_gear_width}
\title{Calculate gear fill in}
\usage{
predict_gear_width(model, coefficient, data)
}
\arguments{
\item{model}{vector of characters defining a model
(see ?linear or ?power)}

\item{coefficient}{coefficient names, must be columns names in data}

\item{data}{a data.frame with the columns, a, b, model, .}
}
\value{
A vector of predicted gear widths.
}
\description{
Predict gear width.
}
\examples{

}
