% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_surface_contact.R
\name{predict_surface_contact}
\alias{predict_surface_contact}
\title{Calculate surface contact}
\usage{
predict_surface_contact(model, fishing_hours, gear_width, fishing_speed)
}
\arguments{
\item{model}{vector of characters defining a model
(see ?surface_contact_models)}

\item{fishing_hours}{the total number of hours fished.}

\item{gear_width}{the average gear width.}

\item{fishing_speed}{the average fishing speed.}
}
\value{
A vector of predicted gear widths.
}
\description{
Predict surface contact.
}
\examples{

}
