% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfdSAR-package.R
\docType{package}
\name{sfdSAR-package}
\alias{sfdSAR-package}
\alias{sfdSAR}
\title{Functions to calculate Swept area ratio and Surface and subsurface
abrasion from VMS data}
\description{
Functions to calculate Swept area ratio and Surface and subsurface
abrasion from VMS data.  The data is expected to conform to the
ICES WGSFD data call format.
}
\references{
ICES Working Group on Spatial Fisheries Data: \url{http://www.ices.dk/community/groups/Pages/WGSFD.aspx}.
}
\author{
Colin Millar, ICES WGSFD.
}
