% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_contact_models.R
\name{surface_contact_models}
\alias{surface_contact_models}
\alias{trawl_contact}
\alias{danish_seine_contact}
\alias{scottish_seine_contact}
\title{Surface contact models}
\usage{
trawl_contact(fishing_hours, gear_width, fishing_speed)

danish_seine_contact(fishing_hours, gear_width, fishing_speed)

scottish_seine_contact(fishing_hours, gear_width, fishing_speed)
}
\arguments{
\item{fishing_hours}{the number of hours of fishing}

\item{gear_width}{(average) gear width in metres}

\item{fishing_speed}{(average) fishing speed in knots}
}
\value{
A vector of predicted gear widths.
}
\description{
Predict the surface contact of a fishing gear
}
\examples{
# compute surface contact for a trawl gear, fishing for 1 hour, with
# a 85 metres trawl width, at 3 knots.
trawl_contact(fishing_hours = 1,
              gear_width = 85,
              fishing_speed = 3)

}
