% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gear_width.R
\name{predict_gear_width}
\alias{predict_gear_width}
\title{Calculate gear width from vessel characteristics}
\usage{
predict_gear_width(model, coefficient, data)
}
\arguments{
\item{model}{vector of characters defining a model
(see ?linear or ?power)}

\item{coefficient}{coefficient names, must be columns names in data}

\item{data}{a data.frame with the columns, a, b, model, .}
}
\value{
A vector of predicted gear widths.
}
\description{
Predict gear width using vessel length or engine size.
}
\examples{
# very simple example of how to apply this helper function
predict_gear_width("power", "avg_aol", data.frame(firstFactor = 1, secondFactor = 1, avg_aol = 1))

# use the dummy vms dataset
data(test_vms)

\donttest{
# get gear widths and metier lookup from ICES DB
library(icesVMS)
metier_lookup <- get_metier_lookup()
gear_widths <- get_benthis_parameters()

# join widths and lookup
library(dplyr)
aux_lookup <-
  gear_widths \%>\%
  right_join(metier_lookup, by = c("benthisMet" = "benthisMetiers"))

# add aux data to vms
vms <-
  aux_lookup \%>\%
  right_join(test_vms, by = c("leMetLevel6" = "LE_MET_level6"))

# calculate the gear width model
vms$gearWidth_model <-
  predict_gear_width(vms$gearModel, vms$gearCoefficient, vms)

}
}
