% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_contact_models.R
\name{surface_contact_models}
\alias{surface_contact_models}
\alias{trawl_contact}
\alias{danish_seine_contact}
\alias{scottish_seine_contact}
\title{Surface contact models}
\usage{
trawl_contact(fishing_hours, gear_width, fishing_speed)

danish_seine_contact(fishing_hours, gear_width, fishing_speed)

scottish_seine_contact(fishing_hours, gear_width, fishing_speed)
}
\arguments{
\item{a}{the a parameter for the model}

\item{b}{the b parameter for the model}

\item{x}{the covariate used in the model, avg_oal (average overall length),
etc.}
}
\value{
A vector of predicted gear widths.
}
\description{
Predict the surface contact of a fishing gear from its benthis classification.
}
\examples{
oal_linear(1, 1, 1)

}
